Isolate render module types
