+++ src/models/render-module/render-module.ts
@@ -1,0 +1,15 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+export interface RenderModule {
+	name: string;
+	description: string;
+	examples: Array<string>;
+
+	sortingRequired: boolean;
+}
